PID implementation finished.
This PID implementation was tested on an "incubator model" environment
(albeit very simple). This is just a barebones implementation. Anyone
may edit this to fit our specific needs.

The float feedback signal and the float output signal is primarily
arbitrary, change this as we see fit. We probably woudldn't need a
setpoint changer anyway since we always want the bot facing where the
line should be.